Refactored ImageModule as Interaction
